pyproject.toml
[project]
name = "agentic-human-guest-simulation-for-ordering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.10"
dependencies = [
    "asyncio",
    "numpy>=1.20",
    "packaging>=23.2",
    "pip>=25.1.1",
    "pycountry>=24.6.1",
    "pydispatcher>=2.0.7",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "pandas>=2.3.0",
    "pydantic-ai>=0.4.3",
    "pydantic-graph>=0.4.3",
    "sanic-ext>=24.12.0",
    "en-core-web-sm",
    "pytest>=8.3.5",
    # Mac-specific alternative dependencies
    "openai>=1.0.0 ; sys_platform == 'darwin'",
    "nltk>=3.8",
    "pandas>=2.3.0",
    "pydantic-ai>=0.4.3",
    "pydantic-graph>=0.4.3",
    "sanic-ext>=24.12.0",
    "en-core-web-sm",
    "scipy>=1.16.1",
    "matplotlib>=3.10.6",
    "seaborn>=0.13.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
